feat(token): implement token creation and validation, update user DTOs

+++ src/main/java/com/glaiss/users/controller/dto/LoginResponse.java
@@ -1,6 +1,6 @@
 package com.glaiss.users.controller.dto;
 
 public record LoginResponse(String accessToken,
-                            Long expiresIn,
+                            Integer expiresIn,
                             String refreshToken) {
 }
